ai-pin/requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.2
blinker==1.9.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
clean-fid==0.1.35
click==8.2.1
clip-anytorch==2.6.0
contourpy==1.3.2
coqpit==0.0.17
coqpit-config==0.2.1
coqui-tts==0.27.0
coqui-tts-trainer==0.3.1
cycler==0.12.1
dateparser==1.1.8
dctorch==0.1.2
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.74.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
imageio==2.37.0
inflect==5.6.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
k-diffusion==0.1.1.post1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mecab-python3==1.0.5
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.6.0
propcache==0.3.2
protobuf==3.19.6
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.7.1
scipy==1.15.3
sentry-sdk==2.34.1
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.22.1
tqdm==4.67.1
trainer==0.0.20
trampoline==0.1.2
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
